+++ src/Settings/Country.tsx
@@ -1,15 +1,15 @@
 import { useContext } from 'react';
 import { observer } from 'mobx-react';
-import { Page, Header } from '@flumens';
+import { Page, Header, Main } from '@flumens';
 import { NavContext } from '@ionic/react';
+import CountriesList from 'common/Components/CountriesList';
 import appModel from 'common/models/app';
-import Main from './Main';
 
 function SelectCountry() {
   const navigate = useContext(NavContext);
 
-  const onSelect = (e: any) => {
-    appModel.attrs.country = e.target.value; // eslint-disable-line no-param-reassign
+  const onSelect = (newCountry: string) => {
+    appModel.attrs.country = newCountry; // eslint-disable-line no-param-reassign
     appModel.save();
     navigate.goBack();
   };
@@ -17,7 +17,9 @@ function SelectCountry() {
   return (
     <Page id="country-select">
       <Header title="Country" />
-      <Main appModel={appModel} onSelect={onSelect} />
+      <Main>
+        <CountriesList onChange={onSelect} />
+      </Main>
     </Page>
   );
 }